:recycle: Rename to walks
